+++ plugins/GLFW/src/VRGLFWInputDevice.cpp
@@ -46,7 +46,7 @@ VRGLFWInputDevice::~VRGLFWInputDevice() {
 void VRGLFWInputDevice::appendNewInputEventsSinceLastCall(VRDataQueue* queue) {
     glfwPollEvents();
 
-    for (int f = 0; f < _events.size(); f++)
+    for (size_t f = 0; f < _events.size(); f++)
     {
     	queue->push(_events[f]);
     }
